Moving the async initialization code from before the main window is created to after in order to try to alleviate issues that could be caused by something silently hanging before the main window is shown and resulting in the app running as a background task rather than an app.
