+++ App.axaml.cs
@@ -2,11 +2,8 @@ using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Data.Core.Plugins;
 using Avalonia.Markup.Xaml;
-using CommunityToolkit.Mvvm.DependencyInjection;
-using CourseEquivalencyDesktop.Services;
 using CourseEquivalencyDesktop.ViewModels.General;
 using CourseEquivalencyDesktop.Views.General;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace CourseEquivalencyDesktop;
 
@@ -17,22 +14,8 @@ public partial class App : Application
         AvaloniaXamlLoader.Load(this);
     }
 
-    public override async void OnFrameworkInitializationCompleted()
+    public override void OnFrameworkInitializationCompleted()
     {
-        // Register all the services needed for the application to run
-        var collection = new ServiceCollection();
-        collection.AddCommonServices();
-
-        // Creates a ServiceProvider containing services from the provided IServiceCollection
-        var services = collection.BuildServiceProvider();
-
-        // Make the services generally available
-        Ioc.Default.ConfigureServices(services);
-
-        // Load the settings file
-        var userSettingsService = Ioc.Default.GetRequiredService<UserSettingsService>();
-        await userSettingsService.LoadSettings();
-
         if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
         {
             // Line below is needed to remove Avalonia data validation.
