+++ Views/General/MainWindow.axaml.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Threading.Tasks;
 using Avalonia.Controls;
-using Avalonia.Interactivity;
 using CommunityToolkit.Mvvm.DependencyInjection;
 using CourseEquivalencyDesktop.Models;
 using CourseEquivalencyDesktop.Services;
@@ -9,7 +8,6 @@ using CourseEquivalencyDesktop.ViewModels.DatabaseSelectionWizard;
 using CourseEquivalencyDesktop.ViewModels.Universities;
 using CourseEquivalencyDesktop.Views.DatabaseSelectionWizard;
 using CourseEquivalencyDesktop.Views.Universities;
-using Microsoft.EntityFrameworkCore;
 using MainWindowViewModel = CourseEquivalencyDesktop.ViewModels.General.MainWindowViewModel;
 
 namespace CourseEquivalencyDesktop.Views.General;
@@ -17,6 +15,7 @@ namespace CourseEquivalencyDesktop.Views.General;
 public partial class MainWindow : Window
 {
     private IDisposable? createUniversityInteractionDisposable;
+    private IDisposable? spawnDatabaseSelectionWizardInteractionDisposable;
 
     public MainWindow()
     {
@@ -26,34 +25,27 @@ public partial class MainWindow : Window
     protected override void OnDataContextChanged(EventArgs e)
     {
         createUniversityInteractionDisposable?.Dispose();
+        spawnDatabaseSelectionWizardInteractionDisposable?.Dispose();
         if (DataContext is MainWindowViewModel vm)
         {
             createUniversityInteractionDisposable = vm.CreateUniversityInteraction.RegisterHandler(SpawnCreateUniversityWindow);
+            spawnDatabaseSelectionWizardInteractionDisposable = vm.SpawnDatabaseSelectionWizardInteraction.RegisterHandler(SpawnDatabaseSelectionWizardWindow);
         }
 
         base.OnDataContextChanged(e);
     }
 
-    protected override async void OnLoaded(RoutedEventArgs e)
+    protected override void OnOpened(EventArgs e)
     {
-        base.OnLoaded(e);
+        base.OnOpened(e);
 
-        if (DataContext is not MainWindowViewModel mainWindowViewModel)
+        if (DataContext is MainWindowViewModel vm)
         {
-            return;
+            vm.InitializationCommand.Execute(vm);
         }
-
-        if (string.IsNullOrEmpty(Ioc.Default.GetRequiredService<UserSettingsService>().DatabaseFilePath))
-        {
-            await SpawnDatabaseSelectionWizardWindow();
-        }
-
-        await Ioc.Default.GetRequiredService<DatabaseService>().Database.MigrateAsync();
-
-        mainWindowViewModel.CompleteLoad();
     }
 
-    private Task SpawnDatabaseSelectionWizardWindow()
+    private Task<bool?> SpawnDatabaseSelectionWizardWindow(bool? _)
     {
         var databaseSelectionWizardViewModel = Ioc.Default.GetRequiredService<DatabaseSelectionWizardViewModel>();
         var databaseSelectionWizardWindow = new DatabaseSelectionWizardWindow
@@ -73,7 +65,7 @@ public partial class MainWindow : Window
             }
         };
 
-        return databaseSelectionWizardWindow.ShowDialog(this);
+        return databaseSelectionWizardWindow.ShowDialog<bool?>(this);
     }
 
     private async Task<University?> SpawnCreateUniversityWindow(int? id)
